Move the test to config package to avoid making the ConfigImpl as public
