+++ kafka-rest/src/main/java/io/confluent/kafkarest/config/ConfigModule.java
@@ -15,6 +15,7 @@
 
 package io.confluent.kafkarest.config;
 
+import com.google.common.annotations.VisibleForTesting;
 import io.confluent.kafka.serializers.subject.strategy.SubjectNameStrategy;
 import io.confluent.kafkarest.KafkaRestConfig;
 import io.confluent.kafkarest.ratelimit.RateLimitBackend;
@@ -229,7 +230,8 @@ public final class ConfigModule extends AbstractBinder {
   @Target({ElementType.TYPE, ElementType.METHOD, ElementType.FIELD, ElementType.PARAMETER})
   public @interface AvroSerializerConfigs {}
 
-  public static final class AvroSerializerConfigsImpl
+  @VisibleForTesting
+  protected static final class AvroSerializerConfigsImpl
       extends AnnotationLiteral<AvroSerializerConfigs> implements AvroSerializerConfigs {}
 
   @Qualifier
@@ -262,7 +264,8 @@ public final class ConfigModule extends AbstractBinder {
   @Target({ElementType.TYPE, ElementType.METHOD, ElementType.FIELD, ElementType.PARAMETER})
   public @interface JsonschemaSerializerConfigs {}
 
-  public static final class JsonschemaSerializerConfigsImpl
+  @VisibleForTesting
+  protected static final class JsonschemaSerializerConfigsImpl
       extends AnnotationLiteral<JsonschemaSerializerConfigs>
       implements JsonschemaSerializerConfigs {}
 
@@ -376,7 +379,8 @@ public final class ConfigModule extends AbstractBinder {
   @Target({ElementType.TYPE, ElementType.METHOD, ElementType.FIELD, ElementType.PARAMETER})
   public @interface ProtobufSerializerConfigs {}
 
-  public static final class ProtobufSerializerConfigsImpl
+  @VisibleForTesting
+  protected static final class ProtobufSerializerConfigsImpl
       extends AnnotationLiteral<ProtobufSerializerConfigs> implements ProtobufSerializerConfigs {}
 
   @Qualifier
